Initial commit for this printable coil obj project
The initial version of this code is designed to create a printable
model of the test coil for a UKy test of the spin transport interface
in the LANL nEDM experiment. It is a double cos theta coil with square
grooves for the wire windings. More versatility should be added for
future use by our colleagues.